Abstracted a bunch of components out of the homepage and added conditional rendering based on if there are any current events
